+++ CHANGELOG.md
@@ -543,6 +543,7 @@ and populate it via a data migration
 - Show sector code and name in the application
 - Budgets do not collect IATI fields or currency as they are set by default
 - Budgets tables do not show IATI fields and only show the financial year
+- Budgets funding type must be the same as the parent activity
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-37...HEAD
 [release-37]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-36...release-37